+++ ShippingCalculator/ShippingZones/Zone2.cs
@@ -7,9 +7,9 @@
             requiresAdditionalFee = true;
         }
 
-        public override void CalculateFee(double price, ref double fee)
+        public override void CalculateFee(decimal price, ref decimal fee)
         {
-            fee = price * 0.12;
+            fee = price * 0.12m;
         }
     }
 }
